export plot function to utility script
